Added NIS Support and UI Improvements
This commit introduces support for Nvidia Image Scaling (NIS) in the in-game menu, including the addition of UI elements for NIS mode and sharpness, and the ability to save these settings. Checks have also been added to verify if DLSS and NIS are supported on the current hardware, with a warning displayed if they are not. The visibility of the DLSS and NIS settings boxes has been altered to collapse when not visible, rather than simply being hidden. Debug messages have been added to display the current upscale mode and NIS sharpness. Lastly, some redundant code related to saving the upscale mode has been removed.